Add files via upload
decided to put this on git. contains stuff I used to get my data in a workable form. data still needs cleaning.

currently building ActualEloPredictorMk2. the plan is to make a prediction on players' elos based on their general stats as well as how long they played on each character and the stats on each character. I will be making predictors for each character based on stats of those who have that character in their top 3 most played, as well as a general stats predictor.