Add files via upload
in elo predictor changed to using 4/5ths of data for training.
have filled out a lot in mk2, am still tuning the general stuff. 
need to tune the classifiers. 

made sure it would work to get the data onto the ana one, havent filled out the ana stuff, have to rename and fill out a lot of the ana stuff. have to make models for the other chars.